Rough and ready -open and 'makedepend' for Heptagon.
The -open <filename> option loads the definitions in the given (source) file
into the symbol table before compilation. The idea is to treat non-standard
code that does not use 'open' directives within programs.

The -M (makedepend) option is complementary, it causes heptc to process a
given file and write a line to stdout:

    <filename> '=' <defined> ':' <required>

where <defined> and <required> are space-delimited lists of symbol.

For example, heptc -M 0023_m.ept gives:

    0023_m.ept=n0023 v0: n0001 limit

meaning: "The 0023_m.ept file defines the symbols n0023 and v0 using the
symbols n0001 limit."

The idea is to run heptc -M on a bunch of files and feed the results through
the tools/map_deps.py script to create a proper depend file for inclusion in
a Makefile.

The calculated dependencies can then be included in a Makefile to compile a
set of source files in the correct order and pass dependencies using -open.
See the example in tools/Makefile-deps.

The implementation is a nasty hack. It does not work properly with records.
Many other features probably do not work properly either. It treats an
unusual case in the context of a project contract.